fully implemented SSE and NEON cases of intrin.hpp; extended the HAL with some basic math functions
