+++ modules/hal/src/precomp.hpp
@@ -42,3 +42,7 @@
 
 #include "opencv2/hal.hpp"
 #include "opencv2/hal/intrin.hpp"
+#include <algorithm>
+#include <cmath>
+#include <cstdlib>
+#include <float.h>
